requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
aiohttp>=3.9.0
pandas>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy<2.0.0
lxml>=4.9.0
html5lib>=1.1
sqlalchemy>=2.0.0

# RAG System Dependencies
# Pin FAISS to the version used when the index was built to avoid binary incompatibility
faiss-cpu==1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
tenacity>=8.2.0
tqdm>=4.66.0